--- .eslintignore
@@ -1,2 +1,0 @@
-*tmp*/
-*tmp*.*
