+++ src/index.js
@@ -1,6 +1,4 @@
-module.exports = {
-	...require('./flast'),
-	...require('./arborist'),
-	...require('./types'),
-	utils: require('./utils'),
-};
+export * from './flast.js';
+export * from './arborist.js';
+export * from './types.js';
+export * from './utils/index.js';
